Create files with correct ownership instead of using 'chown -R'
Recursive chown is slow in Docker because of the copy-on-write
filesystem. In some setups it can be extremely slow.

It is slow even when the files already have the correct owner. The
option '--from=OLD_OWNER' can be used to only modify files owned by a
particular user (if the old owner is unknown, 'find ! -user NEW_OWNER'
can be used), but it's better to create the files with the correct
ownership in the first place.