Remove repetitions of preferred options in affiliation dropdown
Since Symfony 4.4, the preferred choices of a ChoiceType are displayed
at the top and also in the main list (PR 32658).

It really doesn't suite the Affiliation dropdown, because the
preferred options ('No affiliation' and 'Add affiliation...') are
distinct from the main list of existing affiliations, and repeating
them at the start looks very silly.

Remove the repeated choices in the client, because there is no option
to control the ChoiceType's behaviour.