Speed up chown in Docker
Recursive chown is slow in Docker because of the copy-on-write
filesystem.[1]

Avoid chown by creating files with correct ownership where possible.

Where it is not possible (specifically "ADD domjudge.tar.gz" because
the owner in the archive is arbitrary and can't be overridden[2]),
optimise the case where the files already have the correct ownership.

[1] https://github.com/docker/for-linux issue 388
[2] https://github.com/moby/moby issue 35525