cherry-pick: Speed up recursive chown of /domjudge-src in Docker
Recursive chown is slow in Docker because of the copy-on-write
filesystem.[1] We should instead try to create files with correct
ownership from the beginning.

Docker's ADD directive doesn't support overriding the owner of the
extracted files[2], so we copy domjudge.tar.gz (the release archive)
as-is and then extract it manually with correct ownership. (The
archive will be stored in the intermediate layer, but this is actually
much faster and produces a smaller image than `chown -R`.)

We don't bother to do the same for judgehost.tar.gz (which contains
/opt/domjudge/judgehost; it's created via build-judgehost.sh >
judgehost/Dockerfile.build > chroot-and-tar.sh), because that
directory is small.

[1] https://www.github.com/docker/for-linux/issues/388
[2] https://www.github.com/moby/moby/issues/35525