Add configuration option 'show_team_members_publicly'
Default to true.

The implementation is an ugly hack. The old $members field was renamed
to $publicDescription in commit 1844986ef (Use members field as more
generic freeform field, 2022-03-08), so it doesn't make sense to
toggle whether that field is displayed (it's public by definition).
Instead, this option toggles whether the team members of new
registrations are stored in the $publicDescription field or the
$internalComments field.