Configure "Team members" field in nzpc:configure for online-only NZPC contests
We require the "Team members" field to be filled in, so we can
validate the membership of the winning teams. So we set
'team_members_required' to true.

We don't display the team members publicly (it would be fine to
display them if they were optional, but not if they are required). So
we set 'show_team_members_publicly' to false.