Improve detection of existing affiliations during import in nzpc:configure
Previously, detection of existing affiliations was based only on the
shortname. If the shortname was modified (either in the database or in
the imported file), a duplicate affiliation would be created.

To fix this, also check if an affiliation exists with the full name.

Also display warnings if an existing affiliation is found but the
shortname or name are different, to let the user know that the
database and the imported file are inconsistent.