Add configuration option 'require_affiliation_unless_open'
Background: In the NZPC we have an "Open" category which anyone can
compete in, and we have "School" and "Tertiary" categories for
students enrolled in an institution. So teams that don't select an
institution (affiliation) are not eligible for the "School" or
"Tertiary" categories and need to register in the "Open" category.

When this configuration option is enabled, it requires teams to select
an affiliation during self-registration, unless they register in the
"Open" category.

This option only applies if the "Open" category actually exists.