Install Rust compiler in judgehost image
Rust was added to DOMjudge 8.2 in commit de73ea26a (Add rust language
support (#1665), 2022-09-04).

We would like to enable Rust in the NZPC (as an unofficial language).
Backporting the changes to the database (add Rust to the language
table and add the compile script) is too much work, so that will have
to be done by hand through the DOMjudge web interface until we update
to DOMjudge 8.2.