Edit the Rust compile script
Set TMPDIR to the working directory instead of creating a dedicated
subdirectory (it's fine that TMPDIR is not empty, because well-behaved
programs should use mkstemp() or similar to create unique filenames,
which is what rustc does).

This also fixes a few small issues in the previous implementation
(TMPDIR should be absolute; the exit code was 0 even if rustc fails
due to a syntax error; rmdir will fail if the directory is not empty,
e.g. because rustc was killed).

Also add the standard compile script comment.