Update judgehost from Debian "bullseye" to "bookworm"
Static linking is broken on GHC 8.8.4 [1]. The easiest way to upgrade
is to upgrade the OS release (Debian Backports doesn't have newer
versions of ghc or glibc). This way we also get newer versions of the
other languages.

Also need to switch from libcgroup1 (which is no longer available) to
libcgroup2. See also: domjudge-packaging commit 53b67e3b0 (Use
bookworm for building Docker image (Fixes #154), 2023-08-26).

(Not updating the domserver image because the build fails on
"bookworm".)

[1] https://gitlab.haskell.org/ghc/ghc/-/issues/19029