backport: Always change language of .py extension from Python 2 to 3
The extension .py is initially associated with Python 2, and a
migration changes to Python 3. Previously, this migration was skipped
if Python 3 was enabled.

The commit that switched to PyPy also enabled Python by default,
causing the migration to be skipped and leaving .py associated with
Python 2 (which is EOL).

Fix this by running the migration even if Python 3 is enabled.

This isn't relevant for DOMjudge 8 -- see commit c15ee0042 (Remove
Python2 as a default supported language., 2021-03-23).